+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/__init__.py
@@ -6,6 +6,7 @@ from ._enums import KeyCurveName, KeyExportEncryptionAlgorithm, KeyOperation, Ke
 from ._shared.client_base import ApiVersion
 from ._models import (
     DeletedKey,
+    ExternalKey,
     JsonWebKey,
     KeyAttestation,
     KeyProperties,
@@ -21,6 +22,7 @@ from ._client import KeyClient
 __all__ = [
     "ApiVersion",
     "KeyClient",
+    "ExternalKey",
     "JsonWebKey",
     "KeyAttestation",
     "KeyVaultKey",